+++ Items/BituminousCoal.cs
@@ -3,7 +3,7 @@ namespace Core
     public class BituminousCoal : Item
     {
         public override ItemType Type => ItemType.BituminousCoal;
-        public override uint MaxStack => 200_000_000;
+        public override ulong MaxStack => 200_000_000;
         public override UnitType Units => UnitType.Milligram;
         public BituminousCoal(ulong quantity) : base(quantity) { }
         override public float Width => .3f;
