+++ Items/Chalcopyrite.cs
@@ -7,7 +7,7 @@ namespace Core
         public override string ChemicalFormula => "CuFeS₂";
         public override float? SpecificHeat_JoulesPerMgPerDegreeCelsious => .0004f;
         public override float? MeltingPoint_Celsious => 950;
-        public override uint MaxStack => 200_000_000;
+        public override ulong MaxStack => 200_000_000;
         public override UnitType Units => UnitType.Milligram;
         public Chalcopyrite(ulong quantity) : base(quantity) { }
     }
