+++ Items/ClayFurnaceItem.cs
@@ -5,7 +5,7 @@ namespace Core
     public class ClayFurnaceItem : Item
     {
         public override ItemType Type => ItemType.ClayFurnace;
-        public override uint MaxStack => 1;
+        public override ulong MaxStack => 1;
         public override CharacterType? Builds => CharacterType.ClayFurnace;
         private const string name = "Clay furnace";
         public override string Name => name;
