+++ Items/ConveyorItem.cs
@@ -5,7 +5,7 @@ namespace Core
     public class ConveyorItem : Item
     {
         public override ItemType Type => ItemType.Conveyor;
-        public override uint MaxStack => 16;
+        public override ulong MaxStack => 16;
         public override CharacterType? Builds => CharacterType.Conveyor;
         private const string name = "Conveyor";
         public override string Name => name;
