+++ Items/CopperBar.cs
@@ -5,7 +5,7 @@ namespace Core
     public class CopperBar : Item
     {
         public override ItemType Type => ItemType.CopperBar;
-        public override uint MaxStack => 8;
+        public override ulong MaxStack => 8;
         public override Dictionary<ItemType, uint>? Recipe => null;
         private const string name = "Copper bar";
         public override string Name => name;
