+++ Items/DepotItem.cs
@@ -5,7 +5,7 @@ namespace Core
     public class DepotItem : Item
     {
         public override ItemType Type => ItemType.Depot;
-        public override uint MaxStack => 4;
+        public override ulong MaxStack => 4;
         public override CharacterType? Builds => CharacterType.Depot;
         private const string name = "Depot";
         public override string Name => name;
