+++ Items/LigniteCoal.cs
@@ -9,7 +9,7 @@ namespace Core
         public override string Name => name;
         public override string? ChemicalFormula => null;
         public override CombustionProperties? Combustion => properties;
-        public override uint MaxStack => 200_000_000;
+        public override ulong MaxStack => 200_000_000;
         public override UnitType Units => UnitType.Milligram;
         private static CombustionProperties properties = new()
         {
