+++ Items/MineshaftItem.cs
@@ -5,7 +5,7 @@ namespace Core
     public class MineshaftItem : Item
     {
         public override ItemType Type => ItemType.Mineshaft;
-        public override uint MaxStack => 1;
+        public override ulong MaxStack => 1;
         public override CharacterType? Builds => CharacterType.Mineshaft;
         private const string name = "Mine";
         public override string Name => name;
